+++ src/bytesize.c
@@ -103,16 +103,16 @@ unsigned long long int get_amt(char* input) {
 }
 
 /* Converts an integral number between byte sizes */
-double convert_units(char* input, const char* units_from, const char* units_to) {
+long double convert_units(char* input, const char* units_from, const char* units_to) {
   const long int from = get_factor(units_from);
   const long int to = get_factor(units_to);
-  const int amt = get_amt(input);
+  const unsigned long long int amt = get_amt(input);
   log_debug("Amount To Convert      : %d", amt);
   log_debug("Conversion Factor From : %ld", from);
   log_debug("Conversion Factor To   : %ld", to);
 
-  const double factor = (double) from / to;
-  const double conversion = amt * factor;
+  const long double factor = (double) from / to;
+  const long double conversion = amt * factor;
   return conversion;
 }
